docs(README.md): update README with instructions for accessing Kibana and add screenshots
The README file has been updated to include instructions for accessing Kibana. The previous command for retrieving the password has been removed and replaced with a new command. Additionally, two screenshots have been added to the README, one showing the login page and another showing the home page of Kibana.